+++ CHANGELOG.md
@@ -14,6 +14,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### BREAKING CHANGES
 
+## [2.3.0] - 2021-05-05
+
+### Added
+- feat: Validate that routes and mocks files export an array (#101)
+- feat: Validate routes and route variants format (#101)
+- feat: Validate mocks format (#101)
+- feat: Add an alert when mocks contains non existent routeVariants (#101)
+- feat: Ignore mocks, routes and route variants with duplicated ids (#101)
+- feat: Add an alert if mock "from" property is not valid (not correspondent to any mock id) (#101)
+- feat: Add an alert if there are many variants of the same route in the same mock (#101)
+- feat: Support defining http methods in lowercase
+- chore: Support Node.js v16.x
+
+### Changed
+- chore(deps): update dependencies
+
+### Fixed
+- fix: Routes with delay 0 had delay null in plain representation
+
 ## [2.2.0] - 2021-04-14
 
 ### Added